refactor: added currency controller,data save replaced to another method
