+++ src/ExchangeCurrencyProject/ExchangeCurrencyDisplay/Program.cs
@@ -1,18 +1,13 @@
-﻿using NbrbAPI.Models;
-using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Net.Http;
+﻿using System;
 using System.Threading.Tasks;
 
 namespace ExchangeCurrencyDisplay
 {
     class Program
     {
-        static readonly HttpClient client = new HttpClient();
         static bool alive = true;
         static int input;
+        static readonly ExchangeCurrencyController controller = new ExchangeCurrencyController();
         static async Task Main()
         {
             Console.WriteLine("Вас приветствует справочник валют\n");
@@ -30,13 +25,13 @@ namespace ExchangeCurrencyDisplay
                             await GetCurrencyRateAsync();
                             break;
                         case 2:
-                            await GetCurrencyAsync();
+                            await controller.GetCurrencyAsync();
                             break;
                         case 3:
                             alive = false;
                             break;
                         default:
-                            Console.WriteLine("Неккоректное значение");
+                            Console.WriteLine("Неккоректный ввод");
                             break;
                     }
                 }
@@ -47,7 +42,7 @@ namespace ExchangeCurrencyDisplay
             }
         }
         /// <summary>
-        /// Интерфейс получения данных о курсах
+        /// Интерфейс получения данных о курсе
         /// </summary>
         /// <returns></returns>
         static async Task GetCurrencyRateAsync()
@@ -58,64 +53,14 @@ namespace ExchangeCurrencyDisplay
                 Console.WriteLine("Популярные курсы:\n");
                 Console.WriteLine("Доллар США ID: 145");
                 Console.WriteLine("Eвро ID: 292");
-                Console.WriteLine("Российский рубль ID:298");
+                Console.WriteLine("Российский рубль ID:298\n");
 
                 input = Convert.ToInt32(Console.ReadLine());
-
-                await GetRateAsync(input);
+                await controller.GetRateAsync(input);
             }
             catch (Exception ex)
             {
                 Console.WriteLine(ex.Message);
-            }
-        }
-        /// <summary>
-        /// Получить весь список доступных валют в API
-        /// </summary>
-        public static async Task GetCurrencyAsync()
-        {
-            try
-            {
-                HttpResponseMessage response = await client.GetAsync($"https://www.nbrb.by/api/exrates/currencies");
-                string responseMessage = await response.Content.ReadAsStringAsync();
-                response.EnsureSuccessStatusCode();
-
-                var currency_list = JsonConvert.DeserializeObject<Currency[]>(responseMessage);
-                foreach(Currency cur in currency_list)
-                {
-                    Console.WriteLine($"{cur.Cur_Name} ID: {cur.Cur_ID}");
-                }
-            }
-            catch(HttpRequestException ex)
-            {
-                Console.WriteLine(ex.Message);
-            }  
-        }
-        /// <summary>
-        /// Получить данные о конкретном курсе
-        /// </summary>
-        /// <param name="cur_id">ID курса</param>
-        public static async Task GetRateAsync(int cur_id)
-        {
-            var path = @"convert.txt";
-
-            HttpResponseMessage response = await client.GetAsync($"https://www.nbrb.by/api/exrates/rates/ {cur_id}");
-            string responseMessage = await response.Content.ReadAsStringAsync();
-            response.EnsureSuccessStatusCode();
-
-            Rate currency = JsonConvert.DeserializeObject<Rate>(responseMessage);
-            Console.WriteLine($"{currency.Cur_Name} \t Количество у.е - {currency.Cur_Scale} \t Курс по НБРБ - {currency.Cur_OfficialRate} \t Дата обновления - {currency.Date.ToLongDateString()}\n");
-            
-            Console.WriteLine("Желаете сохранить информацию в блокнот? Нажмите Y для подтверждения, любую клавишу для отмены.\n");
-            var key = Console.ReadKey();
-
-            if (key.Key == ConsoleKey.Y)
-            {
-                using (StreamWriter sw = new StreamWriter(path, true, System.Text.Encoding.UTF8))
-                {
-                    await sw.WriteLineAsync($"{currency.Cur_Scale} {currency.Cur_Abbreviation} = {currency.Cur_OfficialRate} BYN - такой курс на {currency.Date.ToLongDateString()}");
-                }
-                Console.WriteLine("\nДанные успешно сохранены в папку приложения. Путь: bin/Debug/netcoreapp3.1/convert.txt");
             }
         }
     }
